Turns out user ids are waaaaaay bigger

+++ src/main/java/com/monits/anviz/model/User.java
@@ -2,17 +2,17 @@ package com.monits.anviz.model;
 
 public class User {
 	
-	private short id;
+	private long id;
 	
 	private String name;
 
-	public User(short id, String name) {
+	public User(long id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public short getId() {
+	public long getId() {
 		return id;
 	}
